# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventarix API Documentation",
        "description": "Inventarix is an inventory management system designed to help organizations efficiently manage items, track transactions, and control user access. The API provides structured endpoints for authentication, category and item management, staff administration, transaction tracking, and personal user settings.",
        "contact": {
            "email": "ahmadzidan1316@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/admin/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticates a admin using email and password. Returns an access token on success",
                "operationId": "d399b456c3c7e7bd223149b9ae2285d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong email or password."
                    }
                }
            }
        },
        "/api/v1/staff/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticates a staff using email and password. Returns an access token on success",
                "operationId": "78cc54cfb496f62f42491e9ce174979b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "staff@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong email or password."
                    }
                }
            }
        },
        "/api/v1/admin/logout": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logs out the currently authenticated admin and revokes their token.",
                "operationId": "6aae892e05ba2b34e4464f7c8a47e309",
                "responses": {
                    "200": {
                        "description": "Logged out successfully."
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/staff/logout": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logs out the currently authenticated staff and revokes their token.",
                "operationId": "6b22a194bf58a2f43988c1f2b8819d74",
                "responses": {
                    "200": {
                        "description": "Logged out successfully."
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Authentication using Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}